Formatted, moved event handlers
