+++ BackGammon/BGPoint.xaml.cs
@@ -4,8 +4,10 @@ using System.Diagnostics;
 using System.IO;
 using System.Linq;
 using System.Runtime.InteropServices.WindowsRuntime;
+using Windows.ApplicationModel.DataTransfer;
 using Windows.Foundation;
 using Windows.Foundation.Collections;
+using Windows.UI;
 using Windows.UI.Xaml;
 using Windows.UI.Xaml.Controls;
 using Windows.UI.Xaml.Controls.Primitives;
@@ -13,6 +15,7 @@ using Windows.UI.Xaml.Data;
 using Windows.UI.Xaml.Input;
 using Windows.UI.Xaml.Media;
 using Windows.UI.Xaml.Navigation;
+using Windows.UI.Xaml.Shapes;
 
 // The User Control item template is documented at https://go.microsoft.com/fwlink/?LinkId=234236
 
@@ -29,9 +32,11 @@ namespace BackGammon
         {
             set
             {
-                this.RenderTransform = new CompositeTransform() { Rotation = (value ? 0.0 : 180.0), CenterX=50, CenterY=200 };                
+                RenderTransform = new CompositeTransform() { Rotation = value ? 0.0 : 180.0, CenterX = 50, CenterY = 200 };
             }
         }
+
+        public int PointNumber { get; set; }
 
         public Brush TriangleFill
         {
@@ -42,21 +47,58 @@ namespace BackGammon
         // Using a DependencyProperty as the backing store for TriangleFill.  This enables animation, styling, binding, etc...
         public static readonly DependencyProperty TriangleFillProperty =
             DependencyProperty.Register("TriangleFill", typeof(Brush), typeof(BGPoint), new PropertyMetadata(new SolidColorBrush(Windows.UI.Colors.Red)));
+
+        private void Rectangle_DragLeave(object sender, DragEventArgs e)
+        {
+            Debug.WriteLine("Rectangle_DragLeave");
+            Rectangle r = (Rectangle)sender;
+            r.Fill = new SolidColorBrush(Colors.Transparent);
+        }
+
+        private void Rectangle_DragEnter(object sender, DragEventArgs e)
+        {
+            Debug.WriteLine("Rectangle_DragEnter drag enter");
+            Rectangle r = (Rectangle)sender;
+
+            r.Opacity = 0.7;
+            r.Fill = new SolidColorBrush(Colors.LightSeaGreen);
+
+            e.AcceptedOperation = DataPackageOperation.Move;
+
+        }
+
+        private async void Rectangle_Drop(object sender, DragEventArgs e)
+        {
+            Debug.WriteLine("Rectangle_Drop");
+
+            if (e.DataView.Contains(StandardDataFormats.Text))
+            {
+                var name = await e.DataView.GetTextAsync();
+
+                if (sender is Rectangle rect)
+                {
+                    int row = Grid.GetRow(rect);
+                    int col = Grid.GetColumn(rect);
+
+                    //if (grid.FindName(name) is Ellipse checker)
+                    //{
+                    //    Grid.SetRow(checker, row);
+                    //    Grid.SetColumn(checker, col);
+
+                    //    if (row == 0)
+                    //    {
+                    //        checker.RenderTransform = new TranslateTransform() { Y = -80 };
+                    //    }
+                    //    else
+                    //    {
+                    //        checker.RenderTransform = new TranslateTransform() { Y = 80 };
+                    //    }
+
+                    //}
+                    rect.Fill = new SolidColorBrush(Colors.Transparent);
+                    e.Handled = true;
+                }
+            }
+        }
     }
-
-    //public class BoolToRotationConverter : IValueConverter
-    //{
-    //    public object Convert(object value, Type targetType, object parameter, string language)
-    //    {
-            
-    //        Debug.WriteLine($"Converter: {value is bool}");
-    //        Debug.WriteLine($"Converter: {(bool)value}");
-    //        return ((bool)value) ? 0.0 : 180.0;
-    //    }
-
-    //    public object ConvertBack(object value, Type targetType, object parameter, string language)
-    //    {
-    //        throw new NotImplementedException();
-    //    }
-    //}
 }
